+++ Source/Alternet.UI/Drawing/RectD.cs
@@ -193,7 +193,7 @@ namespace Alternet.Drawing
         /// or a <see cref='Height'/> less than or equal to 0.
         /// </summary>
         [Browsable(false)]
-        public readonly bool IsEmpty => (width <= 0) || (height <= 0);
+        public readonly bool SizeIsEmpty => (width <= 0) || (height <= 0);
 
         /// <summary>
         /// Tests whether this <see cref='RectD'/> has all properties equal to 0.
@@ -202,8 +202,13 @@ namespace Alternet.Drawing
         public readonly bool IsZero => (width == 0) && (height == 0) && (x == 0) && (y == 0);
 
         /// <summary>
-        /// TopLeft Property - This is a read-only alias for the Point which is at X, Y
-        /// If this is the empty rectangle, the value will be positive infinity, positive infinity.
+        /// Tests whether this <see cref='RectD'/> has all properties equal to 0.
+        /// </summary>
+        [Browsable(false)]
+        public readonly bool IsEmpty => (width == 0) && (height == 0) && (x == 0) && (y == 0);
+
+        /// <summary>
+        /// This is a read-only alias for the point which is at (X, Y).
         /// </summary>
         [Browsable(false)]
         public readonly PointD TopLeft
@@ -215,8 +220,7 @@ namespace Alternet.Drawing
         }
 
         /// <summary>
-        /// TopRight Property - This is a read-only alias for the Point which is at X + Width, Y
-        /// If this is the empty rectangle, the value will be negative infinity, positive infinity.
+        /// Gets the point which is at (X + Width, Y).
         /// </summary>
         [Browsable(false)]
         public readonly PointD TopRight
@@ -228,8 +232,7 @@ namespace Alternet.Drawing
         }
 
         /// <summary>
-        /// BottomLeft Property - This is a read-only alias for the Point which is at X, Y + Height
-        /// If this is the empty rectangle, the value will be positive infinity, negative infinity.
+        /// Gets the point which is at (X, Y + Height).
         /// </summary>
         [Browsable(false)]
         public readonly PointD BottomLeft
@@ -241,9 +244,7 @@ namespace Alternet.Drawing
         }
 
         /// <summary>
-        /// BottomRight Property - This is a read-only alias for the Point
-        /// which is at X + Width, Y + Height
-        /// If this is the empty rectangle, the value will be negative infinity, negative infinity.
+        /// Gets the point which is at (X + Width, Y + Height).
         /// </summary>
         [Browsable(false)]
         public readonly PointD BottomRight
@@ -788,7 +789,8 @@ namespace Alternet.Drawing
         /// <summary>
         /// Creates a string representation of this object based on the format string
         /// and <see cref="IFormatProvider"/> passed in.
-        /// If the provider is null, the <see cref="System.Globalization.CultureInfo.CurrentCulture"/> is used.
+        /// If the provider is null, the
+        /// <see cref="System.Globalization.CultureInfo.CurrentCulture"/> is used.
         /// See the documentation for <see cref="IFormattable"/> for more information.
         /// </summary>
         /// <returns>
@@ -796,7 +798,7 @@ namespace Alternet.Drawing
         /// </returns>
         public readonly string ConvertToString(string format, IFormatProvider provider)
         {
-            if (IsEmpty)
+            if (IsZero)
             {
                 return "Empty";
             }