+++ whatsnew.md
@@ -1,3 +1,28 @@
+# 0.9.516 (2024 March 5)
+
+- Started VListBox control. This is ListBox descendant which is capable to contain huge number of items.
+- Fixed Alternet.UI.Pal library loader.
+- Graphics: FromScreen, DrawRotatedTextI, BlitI, StretchBlitI.
+- Renamed DrawSlicedImage to DrawImageSliced and defined as Graphics extension #115.
+- AppUtils.ExecuteTerminalCommand.
+- Add ConsoleUtils class.
+
+# 0.9.515 (2024 March 4)
+
+- TextureBrush.
+- RectD: Inflate(), Deflate().
+- DrawingUtils: DrawDoubleBorder, DrawSlicedImage.
+- Control: AddVerticalGroupBox, AddHorizontalGroupBox.
+- ComboBox: DefaultImageVerticalOffset, DefaultImageTextDistance, DefaultImageBorderColor, GetItemImageRect.
+- Improved demo layout.
+- New Bitmap contructor with url param: Bitmap(string url).
+- Graphics: DrawImageI, FillRectangleI.
+- Brush.Transparent.
+- Image: AsBrush, Bounds.
+- Control: AddTabControl, new Group method override.
+- Implemented and used converter for GraphicsUnit.
+- Improved ColorComboBox painting.
+
 # 0.9.514 (2024 March 2)
 
 - Added ColorComboBox. You can check how it looks in Button page of ControlSample demo.